modified structure of test directory and simplified some code
