--- diff_match_patch_test/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "dmp_test"
-version = "0.1.0"
-authors = ["chandan <chandan@universe>"]
-edition = "2018"
-
-[dependencies]
-diff_match_patch = { version = "0.1.0", path = "../diff_match_patch" }
-
-
-
